Fix: Use Nullable int while reading PhenotypeSet data

+++ trait_prediction/main/phenotype_set.py
@@ -92,7 +92,7 @@ class PhenotypeSet(Sequence[Phenotype]):
         """
         raw_phenotype_df = pd.read_csv(
             file_path, sep="\t", index_col=0, dtype={"genomeID": str}
-        ).astype("uint8")
+        ).astype("Int64")
         if raw_phenotype_df.index.name != "genomeID":
             raise ValueError("The index of the PhenotypeSet table must be 'genomeID'")
         phenotype_df = raw_phenotype_df.select_dtypes(include="number")
